fix clippy warnings: use clamp, std consts, derive Clone
- Replace manual epsilon clamp with clippy-recommended .clamp() method
- Use std::f32::consts::SQRT_2 instead of magic number
- Remove unnecessary cast (*size as u64 -> *size)
- Derive Clone on AsyncAetherLink and remove manual clone method
- Run cargo fmt for consistent formatting

+++ benches/aether_bench.rs
@@ -50,7 +50,7 @@ fn bench_stream_sizes(c: &mut Criterion) {
         let lba_stream: Vec<u64> = (0..*size).collect();
         let mut kernel = AetherLinkKernel::default();
 
-        group.throughput(Throughput::Elements(*size as u64));
+        group.throughput(Throughput::Elements(*size));
         group.bench_with_input(
             BenchmarkId::from_parameter(size),
             &lba_stream,
